+++ tests/Gedmo/SoftDeleteable/Fixture/Entity/Article.php
@@ -92,4 +92,12 @@ class Article
     {
         $this->comments[] = $comment;
     }
+
+    /**
+     * @return Collection<int, Comment>
+     */
+    public function getComments(): Collection
+    {
+        return $this->comments;
+    }
 }